Initial networking/iterIO code: a kick-only server
I know that packet serialisation is slightly broken somehow, since
passing the result of a packet read to a write produces a different,
wrongly-sized file for my server packet dumps, although not the few
client dumps I tried. But this works, so I figure I might as well
commit it, even if it does print a broken pipe error for every
connection received.